🔨 Split the requests package
